Make Presentation Adapter generic so it can be reused
